pyproject.toml
[tool.poetry]
name = "gamestore-api"
version = "0.1.0"
description = "A Django application to provide resources for a fictional ecommerce"
authors = ["Cácio Lucas <cclucas060901@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.6"
djangorestframework = "^3.14.0"
python-decouple = "^3.7"
psycopg2 = "^2.9.5"
dj-database-url = "^1.2.0"
django-cors-headers = "^3.13.0"
djangorestframework-simplejwt = "^5.2.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
